refactoring file paths + kernel code
