--- nexus-inference/src/conv/mod.rs
@@ -1,5 +1,0 @@
-mod causal1d;
-mod tcn;
-
-pub use causal1d::Causal1dConv;
-pub use tcn::TinyTcn;
