+++ nexus-inference/src/kernel/dot/avx2.rs
