+++ nexus-inference/src/kernel/dot/mod.rs
