+++ nexus-inference/src/kernel/dot/scalar.rs
